+++ src/front/components/ScrollToTop.jsx
@@ -1,26 +1,17 @@
-import { useEffect, useRef } from "react";
-import PropTypes from "prop-types";
+import { useEffect } from "react";
+import { useLocation } from "react-router-dom";
 
-// This component allows the scroll to go to the beginning when changing the view,
-// otherwise it would remain in the position of the previous view. 
-// Investigate more about this React behavior :D 
+// This component scrolls to the top of the page on every route change.
+// Place it inside <Router> so it has access to the location context.
 
-const ScrollToTop = ({ location, children }) => {
-    const prevLocation = useRef(location);
+const ScrollToTop = ({ children }) => {
+    const { pathname } = useLocation();
 
     useEffect(() => {
-        if (location !== prevLocation.current) {
-            window.scrollTo(0, 0);
-        }
-        prevLocation.current = location;
-    }, [location]);
+        window.scrollTo(0, 0);
+    }, [pathname]);
 
     return children;
 };
 
 export default ScrollToTop;
-
-ScrollToTop.propTypes = {
-    location: PropTypes.object,
-    children: PropTypes.any
-};
